requirements.txt
streamlit
wikipedia
requests
llama-cpp-python
